kb/requirements.txt
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
